3: rename http_server to server
