+++ wallet_server/internal/controller/server/server.go
@@ -1,4 +1,4 @@
-package http_server
+package server
 
 import (
 	"context"
